Extract inheritor
