+++ js/ui/collection/ui.collection_widget.base.js
@@ -1007,15 +1007,11 @@ var CollectionWidget = Widget.inherit({
     },
 
     _createItemByTemplate: function(itemTemplate, renderArgs) {
-        return itemTemplate.render(this._getItemTemplateArgs(renderArgs));
-    },
-
-    _getItemTemplateArgs: function(renderArgs) {
-        return {
+        return itemTemplate.render({
             model: renderArgs.itemData,
             container: renderArgs.container,
             index: renderArgs.index
-        };
+        });
     },
 
     _emptyMessageContainer: function() {
